+++ libs/galley-types/test/unit/Test/Galley/Types.hs
@@ -64,3 +64,5 @@ instance Arbitrary FeatureFlags where
       <*> QC.elements [minBound ..]
       <*> QC.elements [minBound ..]
       <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary